Update to components + tailwind 3
