+++ .editorconfig
@@ -1,0 +1,19 @@
+root = true
+
+[*]
+charset = utf-8
+end_of_line = lf
+insert_final_newline = true
+indent_style = space
+indent_size = 4
+trim_trailing_whitespace = true
+
+[*.config.js]
+indent_size = 2
+
+[*.md]
+trim_trailing_whitespace = false
+
+[*.scss]
+indent_size = 2
+
