+++ composer.json
@@ -1,5 +1,6 @@
 {
     "require": {
+        "php": "~8.1",
         "tightenco/jigsaw": "^1.3",
         "torchlight/torchlight-jigsaw": "^0.3.3"
     },
