+++ package.json
@@ -7,12 +7,13 @@
         "prod": "mix --production"
     },
     "devDependencies": {
+        "fast-glob": "^3.2.11",
         "laravel-mix": "^6.0.39",
         "laravel-mix-jigsaw": "^1.5.0",
         "postcss": "^8.2.4",
         "sass": "^1.49.10",
         "sass-loader": "^12.6.0",
-        "tailwindcss": "^0.5.2",
+        "tailwindcss": "^3.0.23",
         "vue-loader": "^15.9.8",
         "vue-template-compiler": "^2.6.10"
     },
