+++ source/_assets/sass/components/_logo.scss
@@ -1,0 +1,5 @@
+
+.logo {
+  color: var(--logo-color, #282224);
+  background-color: var(--logo-bg-color, #EBEBEB);
+}
