+++ source/_assets/sass/components/_stack.scss
@@ -1,11 +1,11 @@
 .stack {
-  @apply .mt-12;
+  @apply mt-12;
   .stack-top {
     img {
       box-shadow: -.5rem -.5rem 0px rgba(0,0,0,.2), -1rem -1rem 0px config('colors.grey-light');
     }
   }
   .lightbox {
-    img { @apply .block; }
+    img { @apply block; }
   }
 }
