+++ source/_assets/sass/page/_home.scss
@@ -1,13 +1,13 @@
 .home {
   h2, h3, h4, h5, h6 {
-    @apply .text-white;
+    @apply text-white;
   }
 
   a.regular {
-    @apply .font-normal .text-white .border-brand-white-darker .border-b .inline-block .leading-none;
+    @apply font-normal text-white border-brand-white-darker border-b inline-block leading-none;
 
     &:hover {
-      @apply .border-none .no-underline;
+      @apply border-none no-underline;
     }
   }
 }
