--- source/_assets/sass/page/_sharp.scss
@@ -1,13 +1,0 @@
-.page-sharp {
-  .heroicon-component-fill {
-    @apply .fill-current text-brand-cyan;
-  }
-
-  .heroicon-component-accent {
-    @apply .fill-current text-white;
-  }
-  .heroicon-shadows { display: none; }
-  .heroicon-outline {
-    @apply .fill-current text-black;
-  }
-}
