--- source/_assets/sass/page/_sicstine.scss
@@ -1,12 +1,0 @@
-.page-sicstine {
-  .heroicon-component-fill {
-    @apply .fill-current text-white;
-  }
-  .heroicon-component-accent {
-    @apply .fill-current text-brand-cyan;
-  }
-  .heroicon-shadows { display: none; }
-  .heroicon-outline {
-    @apply .fill-current text-black;
-  }
-}
