+++ source/_components/section.blade.php
@@ -1,3 +1,8 @@
+@props([
+    // slots
+    'content' => null,
+])
+
 <div class="pb-12 md:pb-32">
     <div class="relative" v-cloak>
         @if($title??false)
@@ -17,8 +22,8 @@
                 </h2>
             </div>
         @endif
-        <div class="{{ $container??'lg:max-w-md xl:max-w-lg' }}" :class="{'md:ml-48':!isMobile}">
-            {{ $slot }}
+        <div class="{{ $content->attributes['class'] ?? 'lg:max-w-md xl:max-w-lg' }}" :class="{'md:ml-48':!isMobile}">
+            {{ $content ?? $slot }}
         </div>
     </div>
 </div>
