+++ source/_layouts/project.blade.php
@@ -1,34 +1,28 @@
-@extends('_layouts.master',[
-    'background'=>'white',
-    'bodyClass'=>"project $page->bodyClass",
-    'headerClass'=>'bg-black text-brand-white'
-])
 
-@section('content')
-    @component('_partials/section', [
-        'stickyTitle'=>[ 'startOffset'=>-130 ],
-        'container'=>'max-w-xl'
-    ])
-        @slot('title')
+<x-layout class="light project {{ $page->bodyClass }}">
+    <x-slot name="head">
+        @if($page->carouselWidth ?? false)
+            <style>
+                @foreach(\Illuminate\Support\Arr::wrap($widths) as $nth => $width)
+                    @if($width !== 'auto')
+                        .carousel:nth-of-type({{ $nth + 1 }}) .VueCarousel-inner {
+                            max-width: {{ $width }} !important;
+                        }
+                    @endif
+                @endforeach
+            </style>
+        @endif
+    </x-slot>
+
+    <x-section :sticky-title="[ 'startOffset'=>-130 ]">
+        <x-slot name="title">
             {{ $page->title }}
-        @endslot
-        <div class="content">
-            <stacker class="mt-6" :breakpoint="768">
-                @yield('projectContent')
-            </stacker>
-        </div>
-    @endcomponent
-@endsection
+        </x-slot>
 
-@section('style')
-    @if($page->carouselWidth ?? false)
-        @php($widths = is_array($page->carouselWidth) ? $page->carouselWidth : [$page->carouselWidth])
-        @foreach($widths as $nth => $width)
-            @if($width !== 'auto')
-                .carousel:nth-of-type({{ $nth + 1 }}) .VueCarousel-inner {
-                    max-width: {{ $width }} !important;
-                }
-            @endif
-        @endforeach
-    @endif
-@endsection
+        <x-slot name="content" class="content max-w-xl">
+            <stacker class="mt-6" :breakpoint="768">
+                @yield('content')
+            </stacker>
+        </x-slot>
+    </x-section>
+</x-layout>
