+++ source/_posts/leveraging-sharp-laravel-open-source-content-management-framework.md
@@ -1,6 +1,6 @@
 ---
 extends: _layouts.post
-section: postContent
+section: content
 ---
 
 # Leveraging Sharp, an open source content management framework, in a Laravel project
