+++ source/_projects/parcs-materiels-grand-est.md
@@ -1,6 +1,6 @@
 ---
 extends: _layouts.project
-section: projectContent
+section: content
 title: Parcs matériels Grand Est
 
 ---
