+++ source/_projects/treto.md
@@ -1,6 +1,6 @@
 ---
 extends: _layouts.project
-section: projectContent
+section: content
 title: Treto
 carouselWidth: ['auto', 30rem]
 ---
