+++ source/_projects/trig.md
@@ -1,6 +1,6 @@
 ---
 extends: _layouts.project
-section: projectContent
+section: content
 title: Trigalière
 carouselWidth: 30rem
 ---
