+++ source/blog.blade.php
@@ -1,0 +1,6 @@
+
+
+
+<x-layout>
+
+</x-layout>
