+++ source/en.blade.php
@@ -1,293 +1,306 @@
-@extends('_layouts.master', [
-    'background' => 'brand-darkblue',
-    'lang' => 'en',
-    'bodyClass' => 'home',
-    'headerClass' => 'text-brand-brown bg-brand-white'
-])
 
-@section('content')
+
+<x-layout class="home" lang="en">
     <div class="text-brand-white">
-        @component('_partials.section', ['large'=>true])
-            @component('_partials.paragraph')
-                @slot('title')
+        <x-section large>
+            <x-paragraph>
+                <x-slot name="title">
                     Quality web development
-                @endslot
+                </x-slot>
                 We build since 2007 websites and mobile apps based on open source technologies,
                 with an "artisan" state of mind: a tailor-made development that allows control at all levels.
                 We have mainly focused on <a href="https://laravel.com" class="regular">Laravel</a>
                 and <a href="https://vuejs.org/" class="regular">Vue.js</a> in recent years.
-            @endcomponent
-            @component('_partials.paragraph')
-                Our customers have one thing in common: they want a solution adapted to a 
-                problem that is not necessarily standard, while benefiting from tools that make 
+            </x-paragraph>
+
+            <x-paragraph>
+                Our customers have one thing in common: they want a solution adapted to a
+                problem that is not necessarily standard, while benefiting from tools that make
                 them autonomous on a daily basis.
-            @endcomponent
-            @component('_partials.paragraph')
-                We are located in Strasbourg, a city on the eastern border of France, 
+            </x-paragraph>
+
+            <x-paragraph>
+                We are located in Strasbourg, a city on the eastern border of France,
                 but we can of course work remotely.
-            @endcomponent
-        @endcomponent
+            </x-paragraph>
+        </x-section>
 
-        @component('_partials.section')
-            @slot('title')
+        <x-section>
+            <x-slot name="title">
                 Our approach
-            @endslot
-            @component('_partials.paragraph')
-                @slot('title')
+            </x-slot>
+
+            <x-paragraph>
+                <x-slot name="title">
                     We attach great<br>importance to the code
-                @endslot
-                Our job is to understand a client's need, translate it into shared 
-                specifications on both sides and ultimately convert it into code: 
+                </x-slot>
+                Our job is to understand a client's need, translate it into shared
+                specifications on both sides and ultimately convert it into code:
                 that's what we focus all our efforts on.
-            @endcomponent
-            @component('_partials.paragraph')
-                Experience shows us that this code, to be reliable and able to adapt to future needs, 
-                must be mastered (which implies to depend only on open source projects), 
-                concise (so that it remains understandable over time, and that it is simple to replace 
-                if a new need is expressed), and systematically accompanied by an automated suite of unit 
-                and functional tests (used at first write stage and even more after, to limit as much as possible regressions 
+            </x-paragraph>
+
+            <x-paragraph>
+                Experience shows us that this code, to be reliable and able to adapt to future needs,
+                must be mastered (which implies to depend only on open source projects),
+                concise (so that it remains understandable over time, and that it is simple to replace
+                if a new need is expressed), and systematically accompanied by an automated suite of unit
+                and functional tests (used at first write stage and even more after, to limit as much as possible regressions
                 over the versions).
-            @endcomponent
-            @component('_partials.paragraph')
-                @slot('title')
+            </x-paragraph>
+
+            <x-paragraph>
+                <x-slot name="title">
                     Our clients are partners
-                @endslot
-                Rather than multiplying small projects and contacts, we seek to create partnerships 
-                with our clients to work with them over the long term, in order to ensure 
+                </x-slot>
+                Rather than multiplying small projects and contacts, we seek to create partnerships
+                with our clients to work with them over the long term, in order to ensure
                 continuous maintenance to our software.
-            @endcomponent
-            @component('_partials.paragraph')
-                @slot('title')
+            </x-paragraph>
+
+            <x-paragraph>
+                <x-slot name="title">
                     We are involved<br>in open source
-                @endslot
-                We massively use open source projects in our work, and we regularly contribute to the 
-                code of these projects when we find some part where an improvement could be made, 
+                </x-slot>
+                We massively use open source projects in our work, and we regularly contribute to the
+                code of these projects when we find some part where an improvement could be made,
                 whether by fixing a bug or adding a new functionality.
-            @endcomponent
-            @component('_partials.paragraph')
+            </x-paragraph>
+
+            <x-paragraph>
                 In the same spirit, we developed and maintain several open source projects, among which we should mention the content management framework Sharp:
-            @endcomponent
+            </x-paragraph>
 
-            @component('_partials.paragraph')
-                @slot('subtitle')
+            <x-project>
+                <x-slot name="title">
                     Sharp for Laravel
-                @endslot
-                @slot('date')
+                </x-slot>
+
+                <x-slot name="date">
                     since 2017
-                @endslot
-                @slot('banner')
-                    ../assets/img/sharp/banner.png
-                @endslot
-                Sharp is a content management framework built for Laravel, a toolset which provides help to build a CMS section in a website. 
-                We use it in many projects. Go to <a href="https://sharp.code16.fr">sharp.code16.fr</a> to learn more.
-            @endcomponent
+                </x-slot>
 
-        @endcomponent
+                <x-slot name="banner" src="assets/img/sharp/banner.png"></x-slot>
+                <x-slot name="website_link" href="https://sharp.code16.fr"></x-slot>
 
-        @component('_partials.section')
-            @slot('title')
+                Sharp est une plateforme de développement d’espaces de gestion de contenu et de données,
+                généraliste et puissante. Nous l’utilisons dans beaucoup de nos projets, pour proposer un outil
+                agréable et performant permettant de mettre en ligne les pages, images, vidéos des sites aussi bien que
+                d’administrer des commandes, des produits ou des inscriptions par exemple.
+            </x-project>
+        </x-section>
+
+        <x-section>
+            <x-slot name="title">
                 E-commerce
-            @endslot
-            @component('_partials.paragraph')
-                @slot('title')
-                    E-commerce and ordering systems
-                @endslot
-                @slot('subtitle')
+            </x-slot>
+
+            <h3 class="mb-4 mt-12 text-xl md:text-3xl lg:text-4xl">
+                E-commerce and ordering systems
+            </h3>
+
+            <x-project>
+                <x-slot name="title">
                     EK France
-                @endslot
-                @slot('banner')
-                    /assets/img/ek/banner.png
-                @endslot
+                </x-slot>
+
+                <x-slot name="banner" src="assets/img/ek/banner.png"></x-slot>
+
                 We have been working for years with EK France, which groups 150+ stores
                 (<a href="https://ambianceetstyles.com">Ambiance & Styles</a> and
                 <a href="https://www.culinarion.com">Culinarion</a>).
-                The main project is the platform tant powers both websites, with a full marketplace: 
+                The main project is the platform tant powers both websites, with a full marketplace:
                 Click & Collect and Ship from Store, online payment with multiple accounts, delivery systems,
-                synchronization of fidelity accounts... The websites also handle gift card orders 
+                synchronization of fidelity accounts... The websites also handle gift card orders
                 and wish list management.
-            @endcomponent
+            </x-project>
 
-            @component('_partials.paragraph')
-                @slot('subtitle')
-                    Cursus, a course ordering website
-                @endslot
-                @slot('date')
+            <x-project>
+                <x-slot name="title">
+                    Projet Cursus
+                </x-slot>
+
+                <x-slot name="date">
                     since 2021
-                @endslot
-                @slot('banner')
-                    ../assets/img/cursus/banner.png
-                @endslot
-                This <a href="https://culturegrandest.fr/cursus/formations/">course ordering system</a> 
-                handles course registration via a complex process depending on various course options and registrant type.
-                <a href="https://sharp.code16.fr">Sharp</a> is leveraged for content and order management. 
-            @endcomponent
+                </x-slot>
 
-            @component('_partials.paragraph')
-                @slot('subtitle')
+                <x-slot name="banner" src="assets/img/cursus/banner.png"></x-slot>
+
+                This <a href="https://culturegrandest.fr/cursus/formations/">course ordering system</a>
+                handles course registration via a complex process depending on various course options and registrant type.
+                <a href="https://sharp.code16.fr">Sharp</a> is leveraged for content and order management.
+            </x-project>
+
+            <x-project>
+                <x-slot name="title">
                     Merkur, a high trafic event shop
-                @endslot
-                @slot('thumbnail')
-                    /assets/img/merkur/thumb.png
-                @endslot
+                </x-slot>
+
+                <x-slot name="thumbnail" src="assets/img/merkur/thumb.png"></x-slot>
+
                 We built, for the Centre international d’Art Verrier de Meisenthal (CIAV),
                 <a href="https://boutique.ciav-meisenthal.fr">an event shop</a> for glass objects
                 including their famous Christmas bulbs.
-            @endcomponent
+            </x-project>
 
-            @component('_partials.paragraph')
-                @slot('subtitle')
+            <x-project>
+                <x-slot name="title">
                     Ekip, a rental service
-                @endslot
-                @slot('thumbnail')
-                    /assets/img/ekip/thumb.png
-                @endslot
+                </x-slot>
+
+                <x-slot name="thumbnail" src="assets/img/ekip/thumb.png"></x-slot>
+
                 We designed and developed <a href="https://parcsmaterielsgrandest.fr">Ekip</a>, a web system for creating
                 advanced quotes and monitoring orders in the context stage equipment rental, from cart to invoice, smoothly bounded with a legacy system.
-            @endcomponent
-        @endcomponent
+            </x-project>
+        </x-section>
 
-        @component('_partials.section')
-            @slot('title')
-                Content websites 
-            @endslot
+        <x-section>
+            <x-slot name="title">
+                Content websites
+            </x-slot>
 
-            @component('_partials.paragraph')
-                @slot('subtitle')
+            <x-project>
+                <x-slot name="title">
                     peugeot-invest.com
-                @endslot
-                @slot('banner')
-                    /assets/img/toshi/banner.png
-                @endslot
+                </x-slot>
+
+                <x-slot name="website_link" href="https://www.peugeot-invest.com/"></x-slot>
+
                 This multilingual website for <a href="https://www.peugeot-invest.com/">Peugeot Invest</a> presents
                 various and numerous content types, including dataviz.
-            @endcomponent
+            </x-project>
 
-            @component('_partials.paragraph')
-                @slot('subtitle')
+            <x-project>
+                <x-slot name="title">
                     jeparticipe.cfdt.fr
-                @endslot
-                @slot('banner')
-                    /assets/img/syzito/banner.jpg
-                @endslot
+                </x-slot>
+
+                <x-slot name="banner" src="assets/img/syzito/banner.jpg"></x-slot>
+
                 <a href="https://jeparticipe.cfdt.fr">Online platform for the CFDT</a> syndicate, with registration / event management.
-            @endcomponent
-            @component('_partials.paragraph')
-                @slot('subtitle')
+            </x-project>
+
+            <x-project>
+                <x-slot name="title">
                     pactedupouvoirdevivre.fr
-                @endslot
-                @slot('banner')
-                    /assets/img/symfonia/banner.jpg
-                @endslot
-                <a href="https://pactedupouvoirdevivre.fr">Main website of the alliance</a> of 60 well known organizations which are major players in the protection 
+                </x-slot>
+
+                <x-slot name="banner" src="assets/img/symfonia/banner.jpg"></x-slot>
+
+                <a href="https://pactedupouvoirdevivre.fr">Main website of the alliance</a> of 60 well known organizations which are major players in the protection
                 of the environment, the fight against poverty, support for migrants, popular education, social and solidarity economy and mutuality.
-            @endcomponent
+            </x-project>
+        </x-section>
 
-        @endcomponent
-
-
-        @component('_partials.section')
-            @slot('title')
+        <x-section>
+            <x-slot name="title">
                 Apps / tools
-            @endslot
+            </x-slot>
 
-            @component('_partials.paragraph')
-                @slot('subtitle')
+            <x-project separator>
+                <x-slot name="title">
                     Treto, a cartographic website
-                @endslot
-                @slot('banner')
-                    /assets/img/treto/banner.png
-                @endslot
+                </x-slot>
+
+                <x-slot name="banner" src="assets/img/treto/banner.png"></x-slot>
+
                 We developed <a href="https://treto.fr/">Treto</a> a regional advanced POI platform
                 for theaters, companies and show producers, with registration and application, account management, validation workflow, ...
-            @endcomponent
+            </x-project>
 
-            @component('_partials.paragraph')
-                @slot('subtitle')
+            <x-project separator>
+                <x-slot name="title">
                     Trig, a cartographic iOS / Android app
-                @endslot
-                @slot('thumbnail')
-                    /assets/img/trig/thumb.png
-                @endslot
+                </x-slot>
+                <x-slot name="thumbnail" src="assets/img/trig/thumb.png"></x-slot>
+
                 The <a href="http://www.domainedelatrigaliere.com/" class="regular">Domaine de la Trigalière</a>,
                 which rents beautiful and luxury vacation houses in a great area made of forests and lakes,
                 asked us to create Trig, a mobile app
                 (<a href="https://itunes.apple.com/fr/app/trigali%C3%A8re/id1332741594?mt=8">iOS</a>
                 et <a href="https://play.google.com/store/apps/details?id=fr.code16.trig">Android</a>)
                 to guide guests across the domain.
-            @endcomponent
+            </x-project>
 
-            @component('_partials.paragraph')
-                @slot('subtitle')
+            <x-project separator>
+                <x-slot name="title">
                     Beatus CRM
-                @endslot
-                @slot('thumbnail')
-                    /assets/img/beatus/thumb.png
-                @endslot
+                </x-slot>
+
+                <x-slot name="thumbnail" src="assets/img/beatus/thumb.png"></x-slot>
+
                 The <a href="https://culturegrandest.fr" class="regular">Agence culturelle Grand Est</a>
                 (the regional entity in charge of cultural development) asked us to work on several projects,
                 linked together by a personalized <em>Single Sign On</em> account system and API synchronizations.
                 First of all, we fully developed a dedicated custom CRM, linked to all other tools
                 to gather data for all contacts.
-            @endcomponent
+            </x-project>
 
-            @component('_partials.paragraph')
-                @slot('subtitle')
+            <x-project>
+                <x-slot name="title">
                     Sapidus
-                @endslot
+                </x-slot>
+
+                <x-slot name="thumbnail" src="assets/img/sapidus/thumb.png"></x-slot>
+
                 We also built an intranet platform, Sapidus, which integrates many internal tools
                 like calendar and resource management, news feed, packages and letters tracking, ...
-            @endcomponent
+            </x-project>
+        </x-section>
 
-        @endcomponent
-            
-
-        @component('_partials.section')
-            @slot('title')
+        <x-section>
+            <x-slot name="title">
                 The team
-            @endslot
-            @component('_partials.paragraph')
-                @slot('subtitle')
+            </x-slot>
+            <x-paragraph>
+                <x-slot name="subtitle">
                     <div class="flex items-center mb-2">
-                        <img class="mr-2 h-12 w-12 rounded-full" src="https://gravatar.com/avatar/{{ md5('philippelonchampt@gmail.com') }}?d=retro" alt="">
+                        <img class="mr-2 h-12 w-12 rounded-full"
+                            src="https://gravatar.com/avatar/{{ md5('philippelonchampt@gmail.com') }}?d=retro" alt="">
                         <span>Philippe Lonchampt</span>
                     </div>
-                @endslot
+                </x-slot>
                 Founder, manager and main developer of several projects.
                 <br><a href="mailto:philippe@code16.fr">email</a>, <a href="https://twitter.com/dvlpp">twitter</a>
-            @endcomponent
-            @component('_partials.paragraph')
-                @slot('subtitle')
+            </x-paragraph>
+
+            <x-paragraph>
+                <x-slot name="subtitle">
                     <div class="flex items-center mb-2">
-                        <img class="mr-2 h-12 w-12 rounded-full" src="https://gravatar.com/avatar/{{ md5('arnaud.becher@gmail.com') }}?d=retro" alt="">
+                        <img class="mr-2 h-12 w-12 rounded-full"
+                            src="https://gravatar.com/avatar/{{ md5('arnaud.becher@gmail.com') }}?d=retro" alt="">
                         <span>Arnaud Becher</span>
                     </div>
-                @endslot
+                </x-slot>
                 Backend developer (PHP and Laravel).
-            @endcomponent
-            @component('_partials.paragraph')
-                @slot('subtitle')
+            </x-paragraph>
+
+            <x-paragraph>
+                <x-slot name="subtitle">
                     <div class="flex items-center mb-2">
-                        <img class="mr-2 h-12 w-12 rounded-full" src="https://gravatar.com/avatar/{{ md5('guingand.antoine@gmail.com') }}?d=retro" alt="">
+                        <img class="mr-2 h-12 w-12 rounded-full"
+                            src="https://gravatar.com/avatar/{{ md5('guingand.antoine@gmail.com') }}?d=retro" alt="">
                         <span>Antoine Guingand</span>
                     </div>
-                @endslot
+                </x-slot>
                 Front-end developer (mobile apps, Vue.js).
-            @endcomponent
-            @component('_partials.paragraph')
+            </x-paragraph>
+
+            <x-paragraph>
                 —<br>
-                Of course, when the project requires it, we surround ourselves with external skills 
+                Of course, when the project requires it, we surround ourselves with external skills
                 (other developers, graphic artists or designers).
-            @endcomponent
-        @endcomponent
+            </x-paragraph>
+        </x-section>
     </div>
 
     <div class="text-brand-white">
-        @component('_partials.section', ['large'=>true])
+        <x-section large>
             Code 16<br>
             <a href="http://www.gothamscm.com/" class="regular">SCM Gotham</a><br>
             24 rue du Vieux marché aux Vins, 67000 Strasbourg, France<br>
             Philippe Lonchampt: <a href="mailto:philippe@code16.fr" class="regular">philippe@code16.fr</a><br>
             <a href="https://github.com/code16" class="regular">Github</a> — <a href="https://medium.com/code16" class="regular">Medium</a>
-        @endcomponent
+        </x-section>
     </div>
-@endsection
+</x-layout>
